Update USA AIRLINE PERFORMANCE
this code will run on a local Jupyter lab, to run on SN lab, make the following adjustments:
Replace this line 

app = JupyterDash(__name__) 

to this line

app=dash.Dash(__name__)

Then comment on this line if exist in your code:

#JupyterDash.infer_jupyter_proxy_config()    

Later at the end replace code like this:

# Run the app
if __name__ == '__main__':
    app.run_server()

Then click again file---> save. (Do not use ctrl+S as it will not work for theia lab).

After that click on terminal---> new terminal and run all the commands in the terminal.